refactor(@angular-devkit/build-webpack): cleanup stable architect API file names
