+++ packages/angular_devkit/build_webpack/builders.json
@@ -2,26 +2,22 @@
   "$schema": "../architect/src/builders-schema.json",
   "builders": {
     "build": {
-      "implementation": "./src/webpack/index2",
-      "class": "./src/webpack",
+      "implementation": "./src/webpack",
       "schema": "./src/webpack/schema.json",
       "description": "Build a webpack app."
     },
     "webpack": {
-      "implementation": "./src/webpack/index2",
-      "class": "./src/webpack",
+      "implementation": "./src/webpack",
       "schema": "./src/webpack/schema.json",
       "description": "Build a webpack app."
     },
     "dev-server": {
-      "implementation": "./src/webpack-dev-server/index2",
-      "class": "./src/webpack-dev-server",
+      "implementation": "./src/webpack-dev-server",
       "schema": "./src/webpack-dev-server/schema.json",
       "description": "Serve a webpack app."
     },
     "webpack-dev-server": {
-      "implementation": "./src/webpack-dev-server/index2",
-      "class": "./src/webpack-dev-server",
+      "implementation": "./src/webpack-dev-server",
       "schema": "./src/webpack-dev-server/schema.json",
       "description": "Serve a webpack app."
     }
